+++ lib/EasyPost/Exception/Api/ExternalApiException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace EasyPost\Exception\Api;
+
+class ExternalApiException extends ApiException
+{
+}