+++ lib/EasyPost/Exception/Api/InternalServerException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace EasyPost\Exception\Api;
+
+class InternalServerException extends ApiException
+{
+}