+++ lib/EasyPost/Exception/Api/NotFoundException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace EasyPost\Exception\Api;
+
+class NotFoundException extends ApiException
+{
+}