+++ lib/EasyPost/Pickup.php
@@ -34,7 +34,6 @@ class Pickup extends EasyPostObject
      * @param array $carriers
      * @param array $services
      * @return Rate
-     * @throws \EasyPost\Exception\Error
      */
     public function lowestRate($carriers = [], $services = [])
     {